+++ src/library/DIPS.Mobile.UI/Components/TextFields/Entry/Entry.cs
@@ -19,9 +19,9 @@ public partial class Entry : Microsoft.Maui.Controls.Entry
 
     private void OnUnfocusedReplaceEmojis(object? sender, FocusEventArgs e)
     {
-        if (!string.IsNullOrEmpty(Text))
+        if (!string.IsNullOrEmpty(Text) && !AllowEmojis)
         {
-            Text = StringFormatter.ReplaceAllEmojisWithPlaceholder(Text);
+            Text = Text.ReplaceAllEmojisWithPlaceholder();
         }
     }